支持仅下载音频，优化get_video

+++ lighting_downloader.py
@@ -13,7 +13,7 @@ from itertools import groupby
 
 
 class Downloader:
-    def __init__(self, videos_dir='videos', sess_data='', video_concurrency=3, part_concurrency=10):
+    def __init__(self, videos_dir='videos', sess_data='', video_concurrency=3, part_concurrency=10, http2=True):
         """
 
         :param videos_dir: 下载到哪个目录，默认当前目录下的为videos中，如果路径不存在将自动创建
@@ -26,7 +26,7 @@ class Downloader:
             os.makedirs(videos_dir)
         cookies = {'SESSDATA': sess_data}
         headers = {'user-agent': 'PostmanRuntime/7.29.0', 'referer': 'https://www.bilibili.com'}
-        self.client = httpx.AsyncClient(headers=headers, cookies=cookies, http2=True)
+        self.client = httpx.AsyncClient(headers=headers, cookies=cookies, http2=http2)
         self.progress = Progress(
             "{task.description}",
             "{task.percentage:>3.0f}%",
@@ -225,13 +225,15 @@ class Downloader:
               for bv in bv_ids]
         )
 
-    async def get_series(self, url: str, quality: int = 0, image=False):
+    async def get_series(self, url: str, quality: int = 0, image=False, subtitle=False, only_audio=False):
         """
         下载某个系列（包括up发布的多p投稿，动画，电视剧，电影等）的所有视频。只有一个视频的情况下仍然可用该方法
 
         :param url: 系列中任意一个视频的url
         :param quality: 下载视频画面的质量，默认0为可下载的最高画质，数字越大质量越低，数值超过范围时默认选取最低画质
         :param image: 是否下载封面
+        :param subtitle: 是否下载字幕 todo
+        :param only_audio: 是否仅下载音频
         :return:
         """
         url = url.split('?')[0]
@@ -243,18 +245,20 @@ class Downloader:
             for idx, i in enumerate(initial_state['videoData']['pages']):
                 p_url = f"{url}?p={idx + 1}"
                 add_name = f"P{idx + 1}-{i['part']}" if len(initial_state['videoData']['pages']) > 1 else ''
-                cors.append(self.get_video(p_url, quality, add_name, image=True if idx == 0 and image else False))
+                cors.append(self.get_video(p_url, quality, add_name,
+                                           image=True if idx == 0 and image else False,
+                                           subtitle=subtitle, only_audio=only_audio))
         elif 'initEpList' in initial_state:  # 动漫，电视剧，电影
             for idx, i in enumerate(initial_state['initEpList']):
                 p_url = i['link']
                 add_name = i['title']
-                cors.append(self.get_video(p_url, quality, add_name, image=image))
+                cors.append(self.get_video(p_url, quality, add_name, image, subtitle, only_audio))
         else:
             rprint(f'[red]未知类型 {url}')
             return
         await asyncio.gather(*cors)
 
-    async def get_video(self, url, quality: int = 0, add_name='', image=False):
+    async def get_video(self, url, quality: int = 0, add_name='', image=False, subtitle=False, only_audio=False):
         """
         下载单个视频
 
@@ -262,6 +266,8 @@ class Downloader:
         :param quality: 下载视频画面的质量，默认0为可下载的最高画质，数字越大质量越低，数值超过范围时默认选取最低画质
         :param add_name: 给文件的额外添加名，用户请直接保持默认
         :param image: 是否下载封面
+        :param subtitle: 是否下载字幕 todo
+        :param only_audio: 是否仅下载音频
         :return:
         """
         await self.sema.acquire()
@@ -277,18 +283,9 @@ class Downloader:
             return
         title = re.search('<h1[^>]*title="([^"]*)"', res.text).groups()[0].strip()
         if add_name:
-            title += f'-{add_name.strip()}'
-        # replace windows illegal character in title
-        title = re.sub(r"[/\\:*?\"<>|]", '', title)
-        if image:
-            img_url = re.search('property="og:image" content="([^"]*)"', res.text).groups()[0]
-            await self.get_img(img_url, title)
-        if os.path.exists(f'{self.videos_dir}/{title}.mp4'):
-            rprint(f'[green]{title}.mp4 已经存在')
-            self.sema.release()
-            return
-        # find video and audio url
-        try:
+            title = f'{title}-{add_name.strip()}'
+        title = re.sub(r"[/\\:*?\"<>|]", '', title)  # replace windows illegal character in title
+        try:  # find video and audio url
             play_info = re.search('<script>window.__playinfo__=({.*})</script><script>', res.text).groups()[0]
             play_info = json.loads(play_info)
             # choose quality
@@ -304,23 +301,60 @@ class Downloader:
             rprint(f'[rgb(234,122,153)]{title} 需要大会员，或该地区不支持')
             self.sema.release()
             return
-        task_id = self.progress.add_task(total=1,
-                                         description=title if len(title) < 43 else f'{title[:20]}...{title[-20:]}')
-        await asyncio.gather(
-            self._get_media(video_urls, f'{title}-video', task_id),
-            self._get_media(audio_urls, f'{title}-audio', task_id)
-        )
-        self.sema.release()  # let next task begin
-        await anyio.run_process(
-            ['ffmpeg', '-i', f'{self.videos_dir}/{title}-video.m4s', '-i', f'{self.videos_dir}/{title}-audio.m4s',
-             '-codec', 'copy', f'{self.videos_dir}/{title}.mp4', '-loglevel', 'quiet'])
-        os.remove(f'{self.videos_dir}/{title}-video.m4s')
-        os.remove(f'{self.videos_dir}/{title}-audio.m4s')
+
+        task_id = self.progress.add_task(
+            total=0,
+            description=title if len(title) < 43 else f'{title[:20]}...{title[-20:]}', visible=False)
+        cors = []
+        # add cor according to params
+        if not only_audio:
+            if os.path.exists(f'{self.videos_dir}/{title}.mp4'):
+                rprint(f'[green]{title}.mp4 已经存在')
+            else:
+                cors.append(self._get_media(video_urls, f'{title}-video', task_id))
+                cors.append(self._get_media(audio_urls, f'{title}-audio', task_id))
+        else:
+            if os.path.exists(f'{self.videos_dir}/{title}.mp3'):
+                rprint(f'[green]{title}.mp3 已经存在')
+            else:
+                cors.append(self._get_media(audio_urls, f'{title}-audio', task_id))
+        if len(cors) > 0:
+            self.progress.update(task_id, total=1, visible=True)
+        # additional task
+        if image:
+            img_url = re.search('property="og:image" content="([^"]*)"', res.text).groups()[0]
+            cors.append(self._get_img(img_url, title))
+        if subtitle:
+            cid = audio_urls[0].split('/')[-2]
+            cors.append(self.get_subtitle(url, cid=cid))
+
+        await asyncio.gather(*cors)
+        self.sema.release()
+
+        if only_audio and not os.path.exists(f'{self.videos_dir}/{title}.mp3'):
+            os.rename(f'{self.videos_dir}/{title}-audio.m4s', f'{self.videos_dir}/{title}.mp3')
+        elif not only_audio and not os.path.exists(f'{self.videos_dir}/{title}.mp4'):
+            await anyio.run_process(
+                ['ffmpeg', '-i', f'{self.videos_dir}/{title}-video.m4s', '-i', f'{self.videos_dir}/{title}-audio.m4s',
+                 '-codec', 'copy', f'{self.videos_dir}/{title}.mp4', '-loglevel', 'quiet'])
+            os.remove(f'{self.videos_dir}/{title}-video.m4s')
+            os.remove(f'{self.videos_dir}/{title}-audio.m4s')
+        else:
+            return
         self.progress.update(task_id, advance=1)
         print(f'{title} 完成')
         self.progress.update(task_id, visible=False)
 
-    async def get_img(self, url, img_name):
+    async def get_subtitle(self, url, cid=None):
+        """todo
+
+        :param url: 视频url
+        :param cid: 如果提供cid则省去一次请求的时间，不提供也没事
+        :return:
+        """
+        pass
+
+    async def _get_img(self, url, img_name):
         if not os.path.exists(f'{self.videos_dir}/imgs'):
             os.makedirs(f'{self.videos_dir}/imgs')
         img_type = url.split('.')[-1]
@@ -379,13 +413,14 @@ class Downloader:
 
 if __name__ == '__main__':
     async def main():
-        d = Downloader(part_concurrency=10, video_concurrency=5)
+        d = Downloader(part_concurrency=10, video_concurrency=10)
         await d.get_series(
-            'https://www.bilibili.com/video/BV1JP4y1K774?p=2'
-            , quality=999, image=True)
+            'https://www.bilibili.com/video/BV1ts411D7mf?spm_id_from=333.999.0.0'
+            , quality=0, image=False, only_audio=True)
         # await d.get_up_videos('18225678', total=5)
         # await d.get_cate_videos('宅舞', order='stow', days=30, keyword='超级敏感', num=100)
-        await d.get_video('https://www.bilibili.com/video/BV1pb4y197Pq?spm_id_from=333.999.0.0', image=True)
+        # await d.get_video('https://www.bilibili.com/video/BV1JP4y1K774?p=2', image=True)
+        # await d.get_video('https://www.bilibili.com/bangumi/play/ep458494?from_spmid=666.25.episode.0', image=True)
         # await d.get_favour('840297609', num=3, series=True)
         # await d.get_collect('630')
         await d.aclose()
